* 'master' of github.com:MStumpp/pse13ss13-app:
  Current work on navi

+++ WalkaRound/WalkaRound-App/src/edu/kit/iti/algo2/pse2013/walkaround/client/model/navigation/output/VisualTextNaviOutput.java
@@ -18,7 +18,7 @@ public class VisualTextNaviOutput implements NaviOutput {
 	}
 
 	@Override
-	public void deliverOutput(double turnAngle, int distToTurn) {
+	public void deliverOutput(double turnAngle, double distToTurn) {
 		
 		
 		// TODO Text setzen
